apps/service/pyproject.toml
[tool.poetry]
name = "allocate-service"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.10"
asgiref = "3.5.2"
cachetools = "5.3.2"
certifi = "2023.11.17"
charset-normalizer = "2.1.1"
coreapi = "2.3.3"
coreschema = "0.0.4"
coverage = "6.5.0"
django = "4.1.3"
django-cors-headers = "3.13.0"
django-extensions = "3.2.1"
django-filter = "22.1"
django-storages = "1.14.2"
djangorestframework = "3.14.0"
djangorestframework-camel-case = "1.3.0"
djangorestframework-simplejwt = "5.2.2"
drf-yasg = "1.21.4"
geographiclib = "2.0"
geopy = "2.3.0"
google-api-core = "2.16.1"
google-auth = "2.27.0"
google-cloud-core = "2.4.1"
google-cloud-storage = "2.14.0"
google-crc32c = "1.5.0"
google-resumable-media = "2.7.0"
googleapis-common-protos = "1.62.0"
gunicorn = "20.1.0"
idna = "3.6"
inflection = "0.5.1"
itypes = "1.2.0"
jinja2 = "3.1.3"
markupsafe = "2.1.4"
packaging = "23.2"
protobuf = "4.25.2"
psycopg2 = "2.9.5"
pyasn1 = "0.5.1"
pyasn1-modules = "0.3.0"
pyjwt = "2.8.0"
pytz = "2022.6"
requests = "2.28.1"
rsa = "4.9"
ruamel-yaml = "0.18.5"
ruamel-yaml-clib = "0.2.8"
slack-sdk = "3.26.2"
sqlparse = "0.4.3"
uritemplate = "4.1.1"
urllib3 = "1.26.18"
flake8 = "^7.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
